+++ dev-packages/browser-integration-tests/suites/tracing/browserTracingIntegration/long-animation-frame-before-navigation/init.js
@@ -9,7 +9,6 @@ Sentry.init({
       enableLongTask: false,
       enableLongAnimationFrame: true,
       instrumentPageLoad: false,
-      enableInp: false,
     }),
   ],
   tracesSampleRate: 1,