+++ dev-packages/browser-integration-tests/suites/tracing/metrics/web-vitals-ttfb/test.ts
@@ -11,14 +11,12 @@ sentryTest('should capture TTFB vital.', async ({ getLocalTestUrl, page }) => {
   const url = await getLocalTestUrl({ testDir: __dirname });
   const eventData = await getFirstSentryEnvelopeRequest<Event>(page, url);
 
-  expect(eventData.measurements).toBeDefined();
-
   // If responseStart === 0, ttfb is not reported
   // This seems to happen somewhat randomly, so we just ignore this in that case
   const responseStart = await page.evaluate("performance.getEntriesByType('navigation')[0].responseStart;");
   if (responseStart !== 0) {
-    expect(eventData.measurements?.ttfb?.value).toBeDefined();
+    expect(eventData.contexts?.trace?.data?.['browser.web_vital.ttfb.value']).toBeGreaterThanOrEqual(0);
   }
 
-  expect(eventData.measurements?.['ttfb.requestTime']?.value).toBeDefined();
+  expect(eventData.contexts?.trace?.data?.['browser.web_vital.ttfb.request_time']).toBeGreaterThanOrEqual(0);
 });