+++ dev-packages/e2e-tests/test-applications/react-router-7-framework-node-20-18/tests/performance/pageload.client.test.ts
@@ -32,7 +32,6 @@ test.describe('client - pageload performance', () => {
       transaction: '/performance',
       type: 'transaction',
       transaction_info: { source: 'route' },
-      measurements: expect.any(Object),
       platform: 'javascript',
       request: {
         url: expect.stringContaining('/performance'),
@@ -85,7 +84,6 @@ test.describe('client - pageload performance', () => {
       transaction: '/performance/with/:param',
       type: 'transaction',
       transaction_info: { source: 'route' },
-      measurements: expect.any(Object),
       platform: 'javascript',
       request: {
         url: expect.stringContaining('/performance/with/sentry'),