+++ packages/browser/src/index.ts
@@ -45,6 +45,7 @@ export { elementTimingIntegration } from '@sentry-internal/browser-utils';
 export { reportPageLoaded } from './tracing/reportPageLoaded';
 export { setActiveSpanInBrowser } from './tracing/setActiveSpan';
 export { spanStreamingIntegration } from './integrations/spanstreaming';
+export { webVitalsIntegration } from './integrations/webVitals';
 
 export type { RequestInstrumentationOptions } from './tracing/request';
 export {